pyproject.toml
[tool.poetry]
name = "003-streamlit-split-and-summarize"
version = "0.1.0"
description = ""
authors = ["Tariq Mansaray <sulaimant.mansaray@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
streamlit = "^1.37.0"
langchain = "^0.2.11"
langchain-openai = "^0.1.19"
python-dotenv = "^1.0.1"
langgraph = "^0.1.19"
langchain-community = "^0.2.10"
pydf = "^12"
chardet = "^5.2.0"
pypdf = "^6.1.1"
beautifulsoup4 = "^4.14.2"
lxml = "^6.0.2"
python-docx = "^1.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"